Added cors to and adjusted tests
