+++ k6/load-tests-heavy.js
@@ -1,14 +1,12 @@
 import { expect } from "https://jslib.k6.io/k6chaijs/4.3.4.3/index.js";
-import { randomString } from "https://jslib.k6.io/k6-utils/1.4.0/index.js";
 import {
-  deleteCredentials,
   getReport,
   getTransactionParams,
-  getTransactions,
+  handleSummary,
+  loadTestOptions as options,
   postCredentials,
   postTransaction,
 } from "./helpers.js";
-import { textSummary } from "https://jslib.k6.io/k6-summary/0.0.1/index.js";
 
 const username = "populated-user";
 const password = "password";
@@ -40,33 +38,7 @@ export const setup = () => {
   }
 };
 
-export const options = {
-  stages: [
-    { duration: "1m", target: 1 }, // Holding
-    { duration: "30s", target: 5 }, // Ramping up
-    { duration: "1m", target: 5 }, // Holding
-    { duration: "30s", target: 10 }, // Ramping up
-    { duration: "1m", target: 10 }, // Holding
-    { duration: "30s", target: 100 }, // Ramping up
-    { duration: "1m", target: 100 }, // Holding
-    { duration: "30s", target: 0 }, // Ramping down
-  ],
-};
-
-export const handleSummary = (data) => {
-  const now = new Date();
-  const timestamp = now.toISOString();
-
-  const summary = textSummary(data, {
-    indent: " ",
-    enableColors: true,
-  });
-
-  return {
-    [`/results/summary_${timestamp}.txt`]: summary,
-    stdout: summary,
-  };
-};
+export { handleSummary, options };
 
 export default () => {
   const reportResponse = getReport(
