+++ rust-budget-api/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 tokio = { version = "1", features = ["full"] }
 axum = { version = "0.7", features = ["macros"] }
-tower-http = { version = "=0.6.1", features = ["fs"] }
+tower-http = { version = "=0.6.1", features = ["fs", "cors"] }
 diesel = { version = "2.2.0", features = [
     "postgres",
     "uuid",
