+++ rust-budget-api/Dockerfile.dev
@@ -1,4 +1,4 @@
-FROM rust
+FROM rust:1.82.0
 
 WORKDIR /usr/src/app
