+++ rust-budget-api/Dockerfile.prod
@@ -1,4 +1,4 @@
-FROM rust AS build-stage
+FROM rust:1.82.0 AS build-stage
 
 WORKDIR /usr/src/app
